moved folder pages from component folder to client folder
