+++ src/client/App.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
-import Home from './components/Pages/Home/HomePage';
-import RegisterPage from './components/Pages/Register/RegisterPage';
-import LoginPage from './components/Pages/Login/LoginPage';
-import ProjectsPage from './components/Pages/Projects/ProjectsPage';
-import ScreenshotsPage from './components/Pages/Screenshots/ScreenshotsPage';
-import UserPage from './components/Pages/User/UserPage';
+import Home from './Pages/Home/HomePage';
+import RegisterPage from './Pages/Register/RegisterPage';
+import LoginPage from './Pages/Login/LoginPage';
+import ProjectsPage from './Pages/Projects/ProjectsPage';
+import ScreenshotsPage from './Pages/Screenshots/ScreenshotsPage';
+import UserPage from './Pages/User/UserPage';
 import { Provider } from './context/AuthContext';
 
 function App() {
