+++ src/client/Pages/Register/RegisterPage.js
